requirements.txt
asgiref==3.5.2
bleach==5.0.1
boto3==1.25.4
botocore==1.28.4
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.4
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
django-storages==1.13.1
django-summernote==0.8.20.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.7.1
idna==3.4
importlib-metadata==5.0.0
jmespath==1.0.1
Pillow==9.3.0
platformdirs==2.5.2
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12
virtualenv==20.14.1
webencodings==0.5.1
whitenoise==6.2.0
zipp==3.9.0
